roomPoolの中身をroomWebSocketに移動。
